requirements.txt
Django==1.7c1
GDAL==1.11.0
Markdown==2.4.1
django-filter==0.7
django-olwidget==0.61.0
djangorestframework==2.3.14
djangorestframework-gis==0.3
psycopg2==2.5.3
wsgiref==0.1.2
